Handle projections on bokeh plots
